allow meteor to be run from subdirectories
Meteor can be run from subdirectories of the root project, as well as the root of the project.  When run in subdirectories, fast-render was not compiling successfully as it was attempting to find `.meteor/packages` somewhere where it doesn't exist.

This is a rough replication of Meteor's app directory identification code, and should successfully determine the correct location to look for the iron-router package, even when run in a subdirectory.