Allow change/remove of other pubs docs to fix #142
Change PublishContext.changed() and PublishContext.removed() to ensure that the
collection exists, and if there isn't a doc with the requested id in the current
publish context, to look for one that was published in another publish context.
If one is found, changed() adds the doc to the current publish context and
removed() does nothing. If there isn't, both throw the same errors that would be
thrown by the merge box.